correct import order, use isEmpty and noop, rename folder hook
